25882590: csrf to auth service
