--- usecase/core_test.go
@@ -1,9 +1,0 @@
-package usecase
-
-import (
-	"testing"
-)
-
-func TestCreateAndKillSession(t *testing.T) {
-
-}
